Fixed VS2010 build warning

+++ opensubdiv/osdutil/topology.cpp
@@ -263,7 +263,7 @@ PxOsdUtilSubdivTopology::ParseFromObjString(
         }
     }
 
-    numVertices = pointPositions->size()/3;
+    numVertices = (int)pointPositions->size()/3;
 
     return true;
 }
